Expand models

+++ portfolio-manager-service/src/main/java/ai/adv/portfoliomanager/model/Portfolio.java
@@ -1,6 +1,7 @@
 package ai.adv.portfoliomanager.model;
 
 import ai.adv.portfoliomanager.model.position.Position;
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
@@ -49,7 +50,7 @@ public class Portfolio {
         .collect(Collectors.toList());
   }
 
-  public Map<String, Integer> getSharesWithNumbers() {
+  public Map<String, BigDecimal> getSharesWithNumbers() {
     return positions.stream()
         .map(Position::getSharesWithNumbers)
         .flatMap(position -> position.entrySet().stream())
